Add a multibranch pipeline for Seed Job
Now we've enabled Spotless for style checking, we don't want to risk
merging PRs without the code style being validated, as then that breaks
our main branch.

We can create a multibranch pipeline for the seed job which runs a
build, which validates the code compiles, and Spotless is happy.